+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # Changelog
 
 ## [Unreleased]
+
+### Added
+
+- Initial Server Setup ([#1](http://github.com/lazoliver/imob-backend/pull/1))